DRY - all logic in one component
