+++ src/Pages/HotPage.jsx
@@ -1,21 +1,10 @@
-import { useSelector, useDispatch } from "react-redux";
-import { MemeComponent } from "../Components/MemeComponent";
-import { toggleFavourite, upvoteMeme, downvoteMeme } from "../store/actions";
+import { MemesPage } from "./MemesPage";
 
 export function HotPage() {
-  const memes = useSelector((state) => state.memes);
-  const dispatch = useDispatch();
-
-  const hotMemes = memes.filter((meme) => meme.upvotes >= 10);
-
   return (
-    <div>
-      <MemeComponent
-        memes={hotMemes}
-        handleUpvote={(id) => dispatch(upvoteMeme(id))}
-        handleDownvote={(id) => dispatch(downvoteMeme(id))}
-        toggleFavourite={(id) => dispatch(toggleFavourite(id))}
-      />
-    </div>
+    <MemesPage
+      filterFn={(meme) => meme.upvotes >= 10}
+      emptyMessage="Brak gorących memów!"
+    />
   );
 }
